requirements-dev.txt
pytest==6.2.4
pytest-flask==1.2.0
pytest-cov==.2.12.1
Flask-DebugToolbar==0.11.0
